+++ src/main/java/dinkplugin/DinkPluginConfig.java
@@ -388,7 +388,9 @@ public interface DinkPluginConfig extends Config {
     @ConfigItem(
         keyName = "collectionNotifMessage",
         name = "Notification Message",
-        description = "The message to be sent through the webhook. Use %USERNAME% to insert your username and %ITEM% for the item",
+        description = "The message to be sent through the webhook.<br/>" +
+            "Use %USERNAME% to insert your username<br/>" +
+            "Use %ITEM% for the item",
         position = 3,
         section = collectionSection
     )
@@ -421,7 +423,8 @@ public interface DinkPluginConfig extends Config {
     @ConfigItem(
         keyName = "petNotifMessage",
         name = "Notification Message",
-        description = "The message to be sent through the webhook. Use %USERNAME% to insert your username",
+        description = "The message to be sent through the webhook.<br/>" +
+            "Use %USERNAME% to insert your username",
         position = 12,
         section = petSection
     )
@@ -476,7 +479,8 @@ public interface DinkPluginConfig extends Config {
     @ConfigItem(
         keyName = "levelMinValue",
         name = "Minimum Skill Level",
-        description = "The minimum skill level required to send a notification. Useful for filtering out low-level notifications",
+        description = "The minimum skill level required to send a notification.<br/>" +
+            "Useful for filtering out low-level notifications",
         position = 24,
         section = levelSection
     )
@@ -487,7 +491,9 @@ public interface DinkPluginConfig extends Config {
     @ConfigItem(
         keyName = "levelNotifMessage",
         name = "Notification Message",
-        description = "The message to be sent through the webhook. Use %USERNAME% to insert your username and %SKILL% to insert the levelled skill(s)",
+        description = "The message to be sent through the webhook.<br/>" +
+            "Use %USERNAME% to insert your username<br/>" +
+            "Use %SKILL% to insert the levelled skill(s)",
         position = 25,
         section = levelSection
     )
@@ -542,7 +548,8 @@ public interface DinkPluginConfig extends Config {
     @ConfigItem(
         keyName = "lootImageMinValue",
         name = "Screenshot Min Value",
-        description = "The minimum combined loot value to send a screenshot. Must have 'Send Image' enabled",
+        description = "The minimum combined loot value to send a screenshot.<br/>" +
+            "Must have 'Send Image' enabled",
         position = 34,
         section = lootSection
     )
@@ -575,7 +582,10 @@ public interface DinkPluginConfig extends Config {
     @ConfigItem(
         keyName = "lootNotifMessage",
         name = "Notification Message",
-        description = "The message to be sent through the webhook. Use %USERNAME% to insert your username, %LOOT% to insert the loot and %SOURCE% to show the source of the loot",
+        description = "The message to be sent through the webhook.<br/>" +
+            "Use %USERNAME% to insert your username<br/>" +
+            "Use %LOOT% to insert the loot<br/>" +
+            "Use %SOURCE% to show the source of the loot",
         position = 37,
         section = lootSection
     )
@@ -619,7 +629,9 @@ public interface DinkPluginConfig extends Config {
     @ConfigItem(
         keyName = "deathNotifMessage",
         name = "Notification Message",
-        description = "The message to be sent through the webhook. Use %USERNAME% to insert your username, %VALUELOST% to insert the GE value of the stuff you lost",
+        description = "The message to be sent through the webhook.<br/>" +
+            "Use %USERNAME% to insert your username<br/>" +
+            "Use %VALUELOST% to insert the GE value of the stuff you lost",
         position = 43,
         section = deathSection
     )
@@ -641,7 +653,10 @@ public interface DinkPluginConfig extends Config {
     @ConfigItem(
         keyName = "deathNotifPvpMessage",
         name = "PvP notification message",
-        description = "The message to be sent through the webhook. Use %PKER% to insert the killer, %USERNAME% to insert your username, %VALUELOST% to insert the GE value of the stuff you lost",
+        description = "The message to be sent through the webhook.<br/>" +
+            "Use %PKER% to insert the killer<br/>" +
+            "Use %USERNAME% to insert your username<br/>" +
+            "Use %VALUELOST% to insert the GE value of the stuff you lost",
         position = 45,
         section = deathSection
     )
@@ -685,7 +700,11 @@ public interface DinkPluginConfig extends Config {
     @ConfigItem(
         keyName = "slayerNotifMessage",
         name = "Notification Message",
-        description = "The message to be sent through the webhook. Use %USERNAME% to insert your username, %TASK% to insert your task, %POINTS% to show how many points you obtained and %TASKCOUNT% to show how many tasks you have completed.",
+        description = "The message to be sent through the webhook.<br/>" +
+            "Use %USERNAME% to insert your username<br/>" +
+            "Use %TASK% to insert your task<br/>" +
+            "Use %POINTS% to show how many points you obtained<br/>" +
+            "Use %TASKCOUNT% to show how many tasks you have completed",
         position = 53,
         section = slayerSection
     )
@@ -718,7 +737,9 @@ public interface DinkPluginConfig extends Config {
     @ConfigItem(
         keyName = "questNotifMessage",
         name = "Notification Message",
-        description = "The message to be sent through the webhook. Use %USERNAME% to insert your username and %QUEST% to insert the quest that you completed",
+        description = "The message to be sent through the webhook.<br/>" +
+            "Use %USERNAME% to insert your username<br/>" +
+            "Use %QUEST% to insert the quest that you completed",
         position = 62,
         section = questSection
     )
@@ -784,7 +805,8 @@ public interface DinkPluginConfig extends Config {
     @ConfigItem(
         keyName = "clueImageMinValue",
         name = "Screenshot Min Value",
-        description = "The minimum combined value of the items to send a screenshot. Must have 'Send Image' enabled",
+        description = "The minimum combined value of the items to send a screenshot.<br/>" +
+            "Must have 'Send Image' enabled",
         position = 75,
         section = clueSection
     )
@@ -795,7 +817,11 @@ public interface DinkPluginConfig extends Config {
     @ConfigItem(
         keyName = "clueNotifMessage",
         name = "Notification Message",
-        description = "The message to be sent through the webhook. Use %USERNAME% to insert your username, %CLUE% to insert the clue type, %LOOT% to show the loot obtained and %COUNT% to insert how many of those clue types you have completed",
+        description = "The message to be sent through the webhook.<br/>" +
+            "Use %USERNAME% to insert your username<br/>" +
+            "Use %CLUE% to insert the clue type<br/>" +
+            "Use %LOOT% to show the loot obtained<br/>" +
+            "Use %COUNT% to insert how many of those clue types you have completed",
         position = 76,
         section = clueSection
     )
@@ -839,7 +865,10 @@ public interface DinkPluginConfig extends Config {
     @ConfigItem(
         keyName = "speedrunPBMessage",
         name = "PB message",
-        description = "%USERNAME% to insert your username, %QUEST% to insert the quest name, %TIME% to insert your new time",
+        description = "The message to be sent through the webhook.<br/>" +
+            "Use %USERNAME% to insert your username<br/> +" +
+            "Use %QUEST% to insert the quest name<br/>" +
+            "Use %TIME% to insert your new time",
         position = 83,
         section = speedrunSection
     )
@@ -850,7 +879,11 @@ public interface DinkPluginConfig extends Config {
     @ConfigItem(
         keyName = "speedrunMessage",
         name = "Notification message",
-        description = "%USERNAME% to insert your username, %QUEST% to insert the quest name, %TIME% to insert your new time, %BEST% to insert your PB",
+        description = "The message to be sent through the webhook.<br/> +" +
+            "Use %USERNAME% to insert your username<br/>" +
+            "Use %QUEST% to insert the quest name<br/>" +
+            "Use %TIME% to insert your new time<br/>" +
+            "Use %BEST% to insert your PB",
         position = 84,
         section = speedrunSection
     )
@@ -916,7 +949,10 @@ public interface DinkPluginConfig extends Config {
     @ConfigItem(
         keyName = "killCountMessage",
         name = "Notification Message",
-        description = "The message to be sent to the webhook. Use %USERNAME% to insert your username, %BOSS% to insert the NPC name, %COUNT% to insert the kill count",
+        description = "The message to be sent to the webhook.<br/>" +
+            "Use %USERNAME% to insert your username<br/>" +
+            "Use %BOSS% to insert the NPC name<br/>" +
+            "Use %COUNT% to insert the kill count",
         position = 95,
         section = killCountSection
     )
@@ -927,7 +963,11 @@ public interface DinkPluginConfig extends Config {
     @ConfigItem(
         keyName = "killCountBestTimeMessage",
         name = "PB Notification Message",
-        description = "The message to be sent to the webhook upon a personal best time. Use %USERNAME% to insert your username, %BOSS% to insert the NPC name, %COUNT% to insert the kill count, %TIME% to insert the completion time",
+        description = "The message to be sent to the webhook upon a personal best time.<br/>" +
+            "Use %USERNAME% to insert your username,<br/>" +
+            "Use %BOSS% to insert the NPC name<br/>" +
+            "Use %COUNT% to insert the kill count<br/>" +
+            "Use %TIME% to insert the completion time",
         position = 96,
         section = killCountSection
     )
@@ -971,7 +1011,10 @@ public interface DinkPluginConfig extends Config {
     @ConfigItem(
         keyName = "combatTaskMessage",
         name = "Notification Message",
-        description = "The message to be sent to the webhook. Use %USERNAME% to insert your username, %TIER% to insert the task tier, %TASK% to insert the task name",
+        description = "The message to be sent to the webhook.<br/>" +
+            "Use %USERNAME% to insert your username<br/>" +
+            "Use %TIER% to insert the task tier<br/>" +
+            "Use %TASK% to insert the task name",
         position = 103,
         section = combatTaskSection
     )
@@ -1015,7 +1058,11 @@ public interface DinkPluginConfig extends Config {
     @ConfigItem(
         keyName = "diaryMessage",
         name = "Notification Message",
-        description = "The message to be sent to the webhook. Use %USERNAME% to insert your username, %DIFFICULTY% to insert the diary difficulty, %AREA% to insert the diary area, %TOTAL% to insert the total diaries completed",
+        description = "The message to be sent to the webhook.<br/>" +
+            "Use %USERNAME% to insert your username<br/>" +
+            "Use %DIFFICULTY% to insert the diary difficulty<br/>" +
+            "Use %AREA% to insert the diary area<br/>" +
+            "Use %TOTAL% to insert the total diaries completed",
         position = 113,
         section = diarySection
     )