+++ src/main/java/dinkplugin/notifiers/CollectionNotifier.java
@@ -8,14 +8,12 @@ import dinkplugin.util.Utils;
 import dinkplugin.notifiers.data.CollectionNotificationData;
 import net.runelite.client.game.ItemManager;
 import org.apache.commons.lang3.StringUtils;
-import org.jetbrains.annotations.VisibleForTesting;
 
 import javax.inject.Inject;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 public class CollectionNotifier extends BaseNotifier {
-    @VisibleForTesting
     static final Pattern COLLECTION_LOG_REGEX = Pattern.compile("New item added to your collection log: (?<itemName>(.*))");
     public static final String ADDITION_WARNING = "Collection notifier will not fire unless you enable the game setting: Collection log - New addition notification";